+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- concord-server: optional rate limit for process start requests;
 - concord-server: ability to assign a policy to a user;
 - ansible: additional validation for `groupVars`.